Rename mod image -> image_folder
