+++ burn-dataset/src/vision/image_folder.rs
